Cytomining (#53)
* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* progress capture

* add the documentation

* feature walkthrough

* feature walkthrough

* progress capture

* progress capture

* progress capture

* progress capture

* chaos commit - will clean up soon c:

* need to clean up + rerun

* segmentations revamped

* post-hoc reassignments

* align labels

* nbconvert

* Ingest data (#41)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* progress catpure

* progress capture

* ingest data

* ingest data

* data ingestion and preprocessing

* data ingestion

* addressing comments

* Run seg featurization (#39)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* run without the reslicing

* run segmentation:

* ready to run all segmentations for all organoids for one patient

* ready to run all segmentations for all organoids for one patient

* fix the profiling

* fix the texturization feature extraction

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* rerun on HPC now that the patient IDs have been added

* rerun on HPC

* data ingestion and preprocessing

* rerun the segs

* rerun the segs

* fixed bash error

* add patient information to scripts

* merge features

* add cleanup

* test with nexflow

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* adding nextflow

* adding nextflow

* fix the featurization

* change the profiling

* run on HPC

* rerun the NF0014 featurization

* granularity

* rerun features to get the proper profiling

* update the resource profiling

* prep for featurization

* add the ability to run nf on HPC

* add the ability to run nf on HPC

* run

* rm slurm files

* rm slurm files

* fix nf

* fix nf

* run each

* attempt 2

* attempt 2

* testing on HPC

* add each nextflow change

* push to HPC:

* attempt run on HPC:

* run all

* merge the profiles

* Update 2.segment_images/child_segmentation.sh


* Update 1.image_quality_control/scripts/0.cp_image_qc.py


* addressing comments

* rebase

---------


* Fix organoid segmentation (#44)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* add segmentation updates

* rerun segmentations

* update segmentations

* addressing comments

* Data Ingestions (#46)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* preprocessing

* rerun on HPC

* push to HPC

* run all

* run all

* process new patient

* process new patient

* Update 2.segment_images/scripts/5.segment_cells_watershed_method.py


* Update 2.segment_images/scripts/5.segment_cells_watershed_method.py


* addressing comments

* Fix organoid segmentation (#44)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* add segmentation updates

* rerun segmentations

* update segmentations

* addressing comments

* Data Ingestions (#46)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* preprocessing

* rerun on HPC

* push to HPC

* run all

* run all

* process new patient

* process new patient

* progress capture

* add the documentation

* feature walkthrough

* Ingest data (#41)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* progress catpure

* progress capture

* ingest data

* ingest data

* data ingestion and preprocessing

* data ingestion

* addressing comments

* Run seg featurization (#39)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* run without the reslicing

* run segmentation:

* ready to run all segmentations for all organoids for one patient

* ready to run all segmentations for all organoids for one patient

* fix the profiling

* fix the texturization feature extraction

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* run all organoids segmentation on HPC

* rerun on HPC now that the patient IDs have been added

* rerun on HPC

* data ingestion and preprocessing

* rerun the segs

* rerun the segs

* fixed bash error

* add patient information to scripts

* merge features

* add cleanup

* test with nexflow

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* testing nf

* adding nextflow

* adding nextflow

* fix the featurization

* change the profiling

* run on HPC

* rerun the NF0014 featurization

* granularity

* rerun features to get the proper profiling

* update the resource profiling

* prep for featurization

* add the ability to run nf on HPC

* add the ability to run nf on HPC

* run

* rm slurm files

* rm slurm files

* fix nf

* fix nf

* run each

* attempt 2

* attempt 2

* testing on HPC

* add each nextflow change

* push to HPC:

* attempt run on HPC:

* run all

* merge the profiles

* Update 2.segment_images/child_segmentation.sh


* Update 1.image_quality_control/scripts/0.cp_image_qc.py


* addressing comments

* rebase

---------


* Fix organoid segmentation (#44)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* add segmentation updates

* rerun segmentations

* update segmentations

* addressing comments

* Data Ingestions (#46)

* Initial commit

* inital commit + preprocessing (#1)

* inital commit

* update paths

* preprocessing

* run on alpine

* change pathing

* HPC

* local update

* addressing comments

* gitignore

* preprocessing

* rerun on HPC

* push to HPC

* run all

* run all

* process new patient

* process new patient

* feature extraction fixes

* redo profiling

* redo the coloc

* rerun coloc

* rerun coloc

* run areashape again

* git add new profilers

* run segs on HPC

* fix the clean up

* fixing slews of bugs

* remove old scripts

* switch to duckdb

* fix the areashape featurization

* fix pathing

* runcellprofiler cleanup

* checking files

* addressing comments

* feature extraction fixes

* redo profiling

* redo the coloc

* rerun coloc

* rerun coloc

* run areashape again

* git add new profilers

* run segs on HPC

* fix the clean up

* fixing slews of bugs

* remove old scripts

* switch to duckdb

* fix the areashape featurization

* fix pathing

* runcellprofiler cleanup

* checking files

* addressing comments

---------

+++ 2.segment_images/scripts/4.post-hoc_mask_refinement.py
@@ -14,7 +14,7 @@
 # ![Segmentation errors](../media/3D_segmentations_correction_events.png)
 #
 
-# In[ ]:
+# In[1]:
 
 
 import argparse
@@ -71,14 +71,14 @@ if not in_notebook:
     patient = args.patient
 else:
     print("Running in a notebook")
-    well_fov = "C2-2"
-    compartment = "organoid"
+    well_fov = "C5-2"
+    compartment = "nuclei"
     patient = "NF0014"
 
 mask_dir = pathlib.Path(f"../../data/{patient}/processed_data/{well_fov}").resolve()
 
 
-# In[ ]:
+# In[3]:
 
 
 if compartment == "nuclei":
@@ -157,6 +157,19 @@ def calculate_overlap(
 
 
 def merge_sets(list_of_sets: list) -> list:
+    """
+    Merge sets in a list of sets if they have any intersection.
+
+    Parameters
+    ----------
+    list_of_sets : list
+        A list of sets to be merged.
+
+    Returns
+    -------
+    list
+        A list of sets after merging those that intersect.
+    """
     for i, set1 in enumerate(list_of_sets):
         for j, set2 in enumerate(list_of_sets):
             if i != j and len(set1.intersection(set2)) > 0:
@@ -188,7 +201,26 @@ def missing_slice_check(
     window_min: int = 0,
     window_max: int = 2,
     interpolated_rows_to_add: List[int] = [],
-):
+) -> List[pd.DataFrame]:
+    """
+    Check for missing slices in the object information DataFrame and add interpolated rows if necessary.
+
+    Parameters
+    ----------
+    object_information_df : pd.DataFrame
+        The DataFrame containing object information with 'z' and 'label' columns.
+    window_min : int, optional
+        The minimum window size for checking missing slices, by default 0
+    window_max : int, optional
+        The maximum window size for checking missing slices, by default 2
+    interpolated_rows_to_add : List[int], optional
+        A list to store rows to be added for interpolation, by default []
+
+    Returns
+    -------
+    List[pd.DataFrame]
+        A list of DataFrames containing rows to be added for interpolation.
+    """
     max_z = object_information_df["z"].max()
     min_z = object_information_df["z"].min()
     if max_z - min_z > 1:
@@ -223,7 +255,26 @@ def add_min_max_boundry_slices(
     global_min_z: int,
     global_max_z: int,
     interpolated_rows_to_add: List[pd.DataFrame] = [],
-):
+) -> List[pd.DataFrame]:
+    """
+    Add slices to the object information DataFrame that are one slice away from the global min and max z slices.
+
+    Parameters
+    ----------
+    object_information_df : pd.DataFrame
+        The DataFrame containing object information with 'z' and 'label' columns.
+    global_min_z : int
+        The global minimum z slice.
+    global_max_z : int
+        The global maximum z slice.
+    interpolated_rows_to_add : List[pd.DataFrame], optional
+        A list to store rows to be added for interpolation, by default []
+
+    Returns
+    -------
+    List[pd.DataFrame]
+        A list of DataFrames containing rows to be added for interpolation at the min and max z slices.
+    """
     # find labels that are 1 slice away from the min or max and extend the label
     for i, row in object_information_df.iterrows():
         # check if the z slice is one away from the min or max (global min and max)
@@ -248,6 +299,21 @@ def add_masks_where_missing(
     new_mask_image: np.ndarray,
     interpolated_rows_to_add_df: pd.DataFrame,
 ) -> np.ndarray:
+    """
+    Add masks to the new mask image where the slices are missing based on the interpolated rows.
+
+    Parameters
+    ----------
+    new_mask_image : np.ndarray
+        The new mask image to which the slices will be added.
+    interpolated_rows_to_add_df : pd.DataFrame
+        The DataFrame containing the rows to be added for interpolation, with columns 'added_z', 'added_new_label', and 'zslice_to_copy'.
+
+    Returns
+    -------
+    np.ndarray
+        The new mask image with the added slices.
+    """
     for slice in interpolated_rows_to_add_df["added_z"].unique():
         # get the rows that correspond to the slice
         tmp_df = interpolated_rows_to_add_df[
@@ -270,7 +336,7 @@ def add_masks_where_missing(
 
 # #### Constants
 
-# In[6]:
+# In[5]:
 
 
 sliding_window_context = 3
@@ -282,13 +348,13 @@ z_slices = [x for x in range(global_min_z, global_max_z)]
 
 # ### Loop through the slices in a sliding window fashion and correct the segmentation
 
-# In[7]:
+# In[6]:
 
 
 new_mask_image = mask.copy()
 
 
-# In[ ]:
+# In[7]:
 
 
 for z in z_slices[: -(sliding_window_context - 1)]:
@@ -409,5 +475,11 @@ for z in z_slices[: -(sliding_window_context - 1)]:
         new_mask_image=new_mask_image,
         interpolated_rows_to_add_df=interpolated_rows_to_add_df,
     )
+    print("writing the mask for z slice", z)
 
+
+# In[8]:
+
+
+if not mask_output_path.exists():
     tifffile.imwrite(mask_output_path, new_mask_image)
